+++ src/tm_devices/commands/afg3k_commands.py
@@ -29,10 +29,18 @@ from .gen_22daqs_afg.system import System
 from .gen_22daqs_afg.trigger import Trigger
 from .gen_33ijgq_afgawg.abort import Abort
 from .gen_33ijgq_afgawg.calibration import Calibration
+from .gen_fsksdy_lpdmsotekscopepcdpomdoafgawgdsa.miscellaneous import Idn, Tst
+from .gen_fsksdy_lpdmsotekscopepcdpomdoafgawgdsa.status_and_error import (
+    Cls,
+    Esr,
+    Opc,
+    Rst,
+    Stb,
+    Wai,
+)
+from .gen_fst7sp_lpdmsotekscopepcmdodpoafgawgdsa.status_and_error import Opt
 from .gen_ft5uww_lpdmsodpomdoafgawgdsa.calibration import Cal
-from .gen_ft5uww_lpdmsodpomdoafgawgdsa.miscellaneous import Idn, Trg, Tst
-from .gen_ft5uww_lpdmsodpomdoafgawgdsa.status_and_error import Cls, Esr, Opc, Rst, Stb, Wai
-from .gen_fteabn_lpdmsomdodpoafgawgdsa.status_and_error import Opt
+from .gen_ft5uww_lpdmsodpomdoafgawgdsa.miscellaneous import Trg
 
 
 # pylint: disable=too-few-public-methods