+++ src/tm_devices/commands/gen_fx54ua_lpdmsodpomdodsa/miscellaneous.py
@@ -13,7 +13,6 @@ Commands and Queries:
     ```
     - *DDT {<Block>|<QString>}
     - *DDT?
-    - *LRN?
     ```
 """
 
@@ -23,30 +22,6 @@ from ..helpers import SCPICmdRead, SCPICmdWrite
 
 if TYPE_CHECKING:
     from tm_devices.drivers.pi.pi_device import PIDevice
-
-
-class Lrn(SCPICmdRead):
-    """The ``*LRN`` command.
-
-    Description:
-        - This query-only command returns the commands that list the instrument settings, allowing
-          you to record or 'learn' the current instrument settings. You can use these commands to
-          return the instrument to the state it was in when you made the ``*LRN?`` query. This
-          command is identical to the SET command.
-
-    Usage:
-        - Using the ``.query()`` method will send the ``*LRN?`` query.
-        - Using the ``.verify(value)`` method will send the ``*LRN?`` query and raise an
-          AssertionError if the returned value does not match ``value``.
-
-    SCPI Syntax:
-        ```
-        - *LRN?
-        ```
-    """
-
-    def __init__(self, device: Optional["PIDevice"] = None, cmd_syntax: str = "*LRN") -> None:
-        super().__init__(device, cmd_syntax)
 
 
 class Ddt(SCPICmdWrite, SCPICmdRead):