+++ src/tm_devices/commands/mdo4kb_commands.py
@@ -67,35 +67,44 @@ from .gen_1nmc1o_msodpomdo.status_and_error import Psc
 from .gen_1nmc1o_msodpomdo.usbdevice import Usbdevice
 from .gen_1nmc1o_msodpomdo.usbtmc import Usbtmc
 from .gen_16x4xq_mdo.search import Search
-from .gen_e6lgg1_lpdmsodpomdo.totaluptime import Totaluptime
-from .gen_e6606z_lpdmsomdodpo.pause import Pause
+from .gen_e6bmgw_lpdmsotekscopepcdpomdo.totaluptime import Totaluptime
+from .gen_e6wozn_lpdmsotekscopepcmdodpo.pause import Pause
 from .gen_fhrp27_msodpomdodsa.curve import Curve
 from .gen_fhrp27_msodpomdodsa.date import Date
 from .gen_fhrp27_msodpomdodsa.mathvar import Mathvar
 from .gen_fhrp27_msodpomdodsa.save_and_recall import Rcl, Sav
+from .gen_fsksdy_lpdmsotekscopepcdpomdoafgawgdsa.miscellaneous import Idn, Tst
+from .gen_fsksdy_lpdmsotekscopepcdpomdoafgawgdsa.status_and_error import (
+    Cls,
+    Esr,
+    Opc,
+    Rst,
+    Stb,
+    Wai,
+)
+from .gen_fst7sp_lpdmsotekscopepcmdodpoafgawgdsa.status_and_error import Opt
 from .gen_ft5uww_lpdmsodpomdoafgawgdsa.calibration import Cal
-from .gen_ft5uww_lpdmsodpomdoafgawgdsa.miscellaneous import Idn, Trg, Tst
-from .gen_ft5uww_lpdmsodpomdoafgawgdsa.status_and_error import Cls, Esr, Opc, Rst, Stb, Wai
-from .gen_fteabn_lpdmsomdodpoafgawgdsa.status_and_error import Opt
-from .gen_fug7nl_lpdmsodpomdoawgdsa.status_and_error import Ese, Sre
-from .gen_fx54ua_lpdmsodpomdodsa.allev import Allev
-from .gen_fx54ua_lpdmsodpomdodsa.busy import Busy
-from .gen_fx54ua_lpdmsodpomdodsa.dese import Dese
-from .gen_fx54ua_lpdmsodpomdodsa.event import Event
-from .gen_fx54ua_lpdmsodpomdodsa.evmsg import Evmsg
-from .gen_fx54ua_lpdmsodpomdodsa.evqty import Evqty
-from .gen_fx54ua_lpdmsodpomdodsa.factory import Factory
-from .gen_fx54ua_lpdmsodpomdodsa.header import Header
-from .gen_fx54ua_lpdmsodpomdodsa.id import Id
-from .gen_fx54ua_lpdmsodpomdodsa.miscellaneous import Ddt, Lrn
+from .gen_ft5uww_lpdmsodpomdoafgawgdsa.miscellaneous import Trg
+from .gen_fu6dog_lpdmsotekscopepcdpomdoawgdsa.status_and_error import Ese, Sre
+from .gen_fx54ua_lpdmsodpomdodsa.miscellaneous import Ddt
 from .gen_fx54ua_lpdmsodpomdodsa.newpass import Newpass
 from .gen_fx54ua_lpdmsodpomdodsa.password import Password
-from .gen_fx54ua_lpdmsodpomdodsa.rem import Rem
-from .gen_fx54ua_lpdmsodpomdodsa.set import Set
-from .gen_fx54ua_lpdmsodpomdodsa.status_and_error import Pud
 from .gen_fx54ua_lpdmsodpomdodsa.teksecure import Teksecure
-from .gen_fx54ua_lpdmsodpomdodsa.verbose import Verbose
-from .gen_fx54ua_lpdmsodpomdodsa.wavfrm import Wavfrm
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.allev import Allev
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.busy import Busy
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.dese import Dese
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.event import Event
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.evmsg import Evmsg
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.evqty import Evqty
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.factory import Factory
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.header import Header
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.id import Id
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.miscellaneous import Lrn
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.rem import Rem
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.set import Set
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.status_and_error import Pud
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.verbose import Verbose
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.wavfrm import Wavfrm
 from .gen_fzn174_lpdmsodpomdodsa.lock import Lock
 from .gen_fzn174_lpdmsodpomdodsa.unlock import Unlock
 from .helpers import DefaultDictPassKeyToFactory